Fixed framerate issues by decoupling rendering from physics and game logic.
Drawing/rendering and game logic now happen on separate threads, so that when drawing slows down collision detection etc. will still proceed at 200 FPS. Changed FPS counters to be displayed on the screen instead of printed to the terminal. FPS (main) is the rate at which game state is updated and input is handled, FPS (render) is the rate at which graphics are drawn to the screen.

Also fixed some little issues with slopes.